// src/interfaces/project.interface.ts
export interface Project {
    id: number;
    project_name: string;
    project_description: string;
    project_code: string;
    project_location: string;
    address: string;
    city: string;
    state: string;
    lga: string;
    project_duration: string;
    start_date: string;
    end_date: string;
    comment: string;
    status: string;
    date_added: string;
    project_supervisor: string;
    supervisor_id: string;
    createdBy: string;
}

 
  